ui: split HookUI class into separate module
Split the `HookUI` (and `NoninteractiveHookUI`) class into separate
module to break the circular import dependency between `apport.ui` and
`apport.report`.
